Fix handling of assert failures in event handlers
This adds `catch` blocks for `AssertFunctionException` whereever `AbortFunctionException` is being caught. There were
only two remaining places, which were in the event handling code.

Also `AssertFunctionException` no longer extends `ParameterException` as it is not necessarily an indicator of parameter
validation. No functional difference as `ParameterExecption` is not caught exlicitly anywhere.