Add Layer.isStampLayer()
This new method acts as a property of layers, stating whether or not thelayer holds stamps. The properties should be
used in situations where the nature as stamps is important, rather than - for example - as a negative filter excluding
tokens.

Most uses of `Token.isStamp()` have been replaced by this new method.