Define a generic method for looking up tokens by layer
This new method is `Zone.getTokensForLayers(Predicate, boolean)`. Unlike `Zone.getTokensOnLayer(Layer)`, this method
allows an arbitrary predicate to be applied to layers in order to decide which tokens to return. It is a gneeralization
of `getTokensOnLayer()` that allows lookup from any number of layers that match the predicate. In practice, this
predicate will be built from layer properties so that the lookup is kept abstract and general.