Include FOW exposure in pathfinding bounds
Since FOW blocks movement for player views, it can "push" optimal paths anywhere based on the exposure boundary. So we
have to account for it when calculating the bounds of cells to be considered for pathfinding.

Also we now clear the FOW blocked moves cache instead of the VBL blocked moves when FOW changes.